# api/openapi.yaml
openapi: 3.1.0
info:
  title: WantMyTime API
  version: 0.3.0
  description: WantMyTime API. Identity, seller links, availability, bookings and offers, payments through Kora (collect, hold, pay out after the session), refunds, no-shows, reviews and MFA-gated operations. Money is integer minor units with an ISO currency; each seller is priced and paid in their country's currency.
servers:
  - url: http://127.0.0.1:8081
components:
  securitySchemes:
    asideSession:
      type: apiKey
      in: cookie
      name: aside_session
security: [{ asideSession: [] }]
paths:
...
  /api/v1/quotes/{id}/checkout:
    post:
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Existing or initialized hosted checkout (one provider reference per quote makes retries idempotent; the time hold is extended before redirect) for the participant's quote; returns price_minor, fee_minor (bank transfer fee paid by the buyer) and total_minor = price_minor + fee_minor }, '401': { description: Verified buyer required }, '409': { description: Quote or seller unavailable }, '503': { description: Provider controls, credentials or approved economics unavailable } }
...
  /api/v1/offers/{id}/checkout:
    post:
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }, { name: Idempotency-Key, in: header, required: true, schema: { type: string } }]
      responses: { '201': { description: Agreed offer time hold for an account or scoped guest buyer }, '503': { description: Simulator and provider checkout are both unavailable } }
...
  /api/v1/webhooks/kora:
    post:
      security: []
      description: Kora notifications, signed with HMAC-SHA256 of the data object (x-korapay-signature). Charges are queued and verified independently; refund and transfer notices bring the next status check forward.
      responses: { '200': { description: Signed provider event durably accepted or duplicate acknowledged }, '400': { description: Invalid provider event }, '401': { description: Invalid provider signature }, '503': { description: Event not acknowledged because durable storage or controls are unavailable } }
...